+++ src/client/mod.rs
@@ -7,7 +7,7 @@ use std::{borrow::Cow, fmt::Debug, io};
 
 use async_trait::async_trait;
 
-use crate::{error::unexpected_rsp_code_panic_msg, frame::*, slave::*, Result};
+use crate::{frame::*, slave::*, ResponseError, ResponseResult, Result};
 
 #[cfg(feature = "rtu")]
 pub mod rtu;
@@ -22,7 +22,7 @@ pub mod sync;
 #[async_trait]
 pub trait Client: SlaveContext + Send + Debug {
     /// Invoke a Modbus function
-    async fn call(&mut self, request: Request<'_>) -> Result<Response>;
+    async fn call(&mut self, request: Request<'_>) -> ResponseResult;
 }
 
 /// Asynchronous Modbus reader
@@ -112,7 +112,7 @@ impl From<Context> for Box<dyn Client> {
 
 #[async_trait]
 impl Client for Context {
-    async fn call(&mut self, request: Request<'_>) -> Result<Response> {
+    async fn call(&mut self, request: Request<'_>) -> ResponseResult {
         self.client.call(request).await
     }
 }
@@ -129,25 +129,17 @@ impl Reader for Context {
         self.client
             .call(Request::ReadCoils(addr, cnt))
             .await
-            .map(|modbus_rsp| {
-                modbus_rsp.map(|rsp| match rsp {
-                    Response::ReadCoils(mut coils) => {
-                        debug_assert!(coils.len() >= cnt.into());
-                        coils.truncate(cnt.into());
-                        coils
-                    }
-                    others => {
-                        // NOTE: A call to `Client::call` implementation *MUST* always return the `Response` variant matching the `Request` one.
-                        // TIPS: This can be ensured via a call to `verify_response_header`( in 'src/service/mod.rs') before returning from `Client::call`.
-                        unreachable!(
-                            "{}",
-                            unexpected_rsp_code_panic_msg(
-                                FunctionCode::ReadCoils,
-                                others.function_code()
-                            ),
-                        )
-                    }
-                })
+            .map(|result| {
+                result
+                    .map_err(Into::into)
+                    .and_then(|response| match response {
+                        Response::ReadCoils(mut coils) => {
+                            debug_assert!(coils.len() >= cnt.into());
+                            coils.truncate(cnt.into());
+                            Ok(coils)
+                        }
+                        response => Err(ResponseError::UnexpectedResponse { response }),
+                    })
             })
     }
 
@@ -159,25 +151,17 @@ impl Reader for Context {
         self.client
             .call(Request::ReadDiscreteInputs(addr, cnt))
             .await
-            .map(|modbus_rsp| {
-                modbus_rsp.map(|rsp| match rsp {
-                    Response::ReadDiscreteInputs(mut coils) => {
-                        debug_assert!(coils.len() >= cnt.into());
-                        coils.truncate(cnt.into());
-                        coils
-                    }
-                    others => {
-                        // NOTE: A call to `Client::call` implementation *MUST* always return the `Response` variant matching the `Request` one.
-                        // TIPS: This can be ensured via a call to `verify_response_header`( in 'src/service/mod.rs') before returning from `Client::call`.
-                        unreachable!(
-                            "{}",
-                            unexpected_rsp_code_panic_msg(
-                                FunctionCode::ReadDiscreteInputs,
-                                others.function_code()
-                            ),
-                        )
-                    }
-                })
+            .map(|result| {
+                result
+                    .map_err(Into::into)
+                    .and_then(|response| match response {
+                        Response::ReadDiscreteInputs(mut coils) => {
+                            debug_assert!(coils.len() >= cnt.into());
+                            coils.truncate(cnt.into());
+                            Ok(coils)
+                        }
+                        response => Err(ResponseError::UnexpectedResponse { response }),
+                    })
             })
     }
 
@@ -189,24 +173,16 @@ impl Reader for Context {
         self.client
             .call(Request::ReadInputRegisters(addr, cnt))
             .await
-            .map(|modbus_rsp| {
-                modbus_rsp.map(|rsp| match rsp {
-                    Response::ReadInputRegisters(words) => {
-                        debug_assert_eq!(words.len(), cnt.into());
-                        words
-                    }
-                    others => {
-                        // NOTE: A call to `Client::call` implementation *MUST* always return the `Response` variant matching the `Request` one.
-                        // TIPS: This can be ensured via a call to `verify_response_header`( in 'src/service/mod.rs') before returning from `Client::call`.
-                        unreachable!(
-                            "{}",
-                            unexpected_rsp_code_panic_msg(
-                                FunctionCode::ReadInputRegisters,
-                                others.function_code()
-                            ),
-                        )
-                    }
-                })
+            .map(|result| {
+                result
+                    .map_err(Into::into)
+                    .and_then(|response| match response {
+                        Response::ReadInputRegisters(words) => {
+                            debug_assert_eq!(words.len(), cnt.into());
+                            Ok(words)
+                        }
+                        response => Err(ResponseError::UnexpectedResponse { response }),
+                    })
             })
     }
 
@@ -218,24 +194,16 @@ impl Reader for Context {
         self.client
             .call(Request::ReadHoldingRegisters(addr, cnt))
             .await
-            .map(|modbus_rsp| {
-                modbus_rsp.map(|rsp| match rsp {
-                    Response::ReadHoldingRegisters(words) => {
-                        debug_assert_eq!(words.len(), cnt.into());
-                        words
-                    }
-                    others => {
-                        // NOTE: A call to `Client::call` implementation *MUST* always return the `Response` variant matching the `Request` one.
-                        // TIPS: This can be ensured via a call to `verify_response_header`( in 'src/service/mod.rs') before returning from `Client::call`.
-                        unreachable!(
-                            "{}",
-                            unexpected_rsp_code_panic_msg(
-                                FunctionCode::ReadHoldingRegisters,
-                                others.function_code()
-                            ),
-                        )
-                    }
-                })
+            .map(|result| {
+                result
+                    .map_err(Into::into)
+                    .and_then(|response| match response {
+                        Response::ReadHoldingRegisters(words) => {
+                            debug_assert_eq!(words.len(), cnt.into());
+                            Ok(words)
+                        }
+                        response => Err(ResponseError::UnexpectedResponse { response }),
+                    })
             })
     }
 
@@ -254,24 +222,16 @@ impl Reader for Context {
                 Cow::Borrowed(write_data),
             ))
             .await
-            .map(|modbus_rsp| {
-                modbus_rsp.map(|rsp| match rsp {
-                    Response::ReadWriteMultipleRegisters(words) => {
-                        debug_assert_eq!(words.len(), read_count.into());
-                        words
-                    }
-                    others => {
-                        // NOTE: A call to `Client::call` implementation *MUST* always return the `Response` variant matching the `Request` one.
-                        // TIPS: This can be ensured via a call to `verify_response_header`( in 'src/service/mod.rs') before returning from `Client::call`.
-                        unreachable!(
-                            "{}",
-                            unexpected_rsp_code_panic_msg(
-                                FunctionCode::ReadWriteMultipleRegisters,
-                                others.function_code()
-                            ),
-                        )
-                    }
-                })
+            .map(|result| {
+                result
+                    .map_err(Into::into)
+                    .and_then(|response| match response {
+                        Response::ReadWriteMultipleRegisters(words) => {
+                            debug_assert_eq!(words.len(), read_count.into());
+                            Ok(words)
+                        }
+                        response => Err(ResponseError::UnexpectedResponse { response }),
+                    })
             })
     }
 }
@@ -282,51 +242,36 @@ impl Writer for Context {
         self.client
             .call(Request::WriteSingleCoil(addr, coil))
             .await
-            .map(|modbus_rsp| {
-                modbus_rsp.map(|rsp| match rsp {
-                    Response::WriteSingleCoil(rsp_addr, rsp_coil) => {
-                        debug_assert_eq!(addr, rsp_addr);
-                        debug_assert_eq!(coil, rsp_coil);
-                    }
-                    others => {
-                        // NOTE: A call to `Client::call` implementation *MUST* always return the `Response` variant matching the `Request` one.
-                        // TIPS: This can be ensured via a call to `verify_response_header`( in 'src/service/mod.rs') before returning from `Client::call`.
-                        unreachable!(
-                            "{}",
-                            unexpected_rsp_code_panic_msg(
-                                FunctionCode::WriteSingleCoil,
-                                others.function_code()
-                            ),
-                        )
-                    }
-                })
+            .map(|result| {
+                result
+                    .map_err(Into::into)
+                    .and_then(|response| match response {
+                        Response::WriteSingleCoil(rsp_addr, rsp_coil) => {
+                            debug_assert_eq!(addr, rsp_addr);
+                            debug_assert_eq!(coil, rsp_coil);
+                            Ok(())
+                        }
+                        response => Err(ResponseError::UnexpectedResponse { response }),
+                    })
             })
     }
 
     async fn write_multiple_coils<'a>(&'a mut self, addr: Address, coils: &[Coil]) -> Result<()> {
         let cnt = coils.len();
-
         self.client
             .call(Request::WriteMultipleCoils(addr, Cow::Borrowed(coils)))
             .await
-            .map(|modbus_rsp| {
-                modbus_rsp.map(|rsp| match rsp {
-                    Response::WriteMultipleCoils(rsp_addr, rsp_cnt) => {
-                        debug_assert_eq!(addr, rsp_addr);
-                        debug_assert_eq!(cnt, rsp_cnt.into());
-                    }
-                    others => {
-                        // NOTE: A call to `Client::call` implementation *MUST* always return the `Response` variant matching the `Request` one.
-                        // TIPS: This can be ensured via a call to `verify_response_header`( in 'src/service/mod.rs') before returning from `Client::call`.
-                        unreachable!(
-                            "{}",
-                            unexpected_rsp_code_panic_msg(
-                                FunctionCode::WriteMultipleCoils,
-                                others.function_code()
-                            ),
-                        )
-                    }
-                })
+            .map(|result| {
+                result
+                    .map_err(Into::into)
+                    .and_then(|response| match response {
+                        Response::WriteMultipleCoils(rsp_addr, rsp_cnt) => {
+                            debug_assert_eq!(addr, rsp_addr);
+                            debug_assert_eq!(cnt, rsp_cnt.into());
+                            Ok(())
+                        }
+                        response => Err(ResponseError::UnexpectedResponse { response }),
+                    })
             })
     }
 
@@ -334,24 +279,17 @@ impl Writer for Context {
         self.client
             .call(Request::WriteSingleRegister(addr, word))
             .await
-            .map(|modbus_rsp| {
-                modbus_rsp.map(|rsp| match rsp {
-                    Response::WriteSingleRegister(rsp_addr, rsp_word) => {
-                        debug_assert_eq!(addr, rsp_addr);
-                        debug_assert_eq!(word, rsp_word);
-                    }
-                    others => {
-                        // NOTE: A call to `Client::call` implementation *MUST* always return the `Response` variant matching the `Request` one.
-                        // TIPS: This can be ensured via a call to `verify_response_header`( in 'src/service/mod.rs') before returning from `Client::call`.
-                        unreachable!(
-                            "{}",
-                            unexpected_rsp_code_panic_msg(
-                                FunctionCode::WriteSingleRegister,
-                                others.function_code()
-                            ),
-                        )
-                    }
-                })
+            .map(|result| {
+                result
+                    .map_err(Into::into)
+                    .and_then(|response| match response {
+                        Response::WriteSingleRegister(rsp_addr, rsp_word) => {
+                            debug_assert_eq!(addr, rsp_addr);
+                            debug_assert_eq!(word, rsp_word);
+                            Ok(())
+                        }
+                        response => Err(ResponseError::UnexpectedResponse { response }),
+                    })
             })
     }
 
@@ -361,28 +299,20 @@ impl Writer for Context {
         data: &[Word],
     ) -> Result<()> {
         let cnt = data.len();
-
         self.client
             .call(Request::WriteMultipleRegisters(addr, Cow::Borrowed(data)))
             .await
-            .map(|modbus_rsp| {
-                modbus_rsp.map(|rsp| match rsp {
-                    Response::WriteMultipleRegisters(rsp_addr, rsp_cnt) => {
-                        debug_assert_eq!(addr, rsp_addr);
-                        debug_assert_eq!(cnt, rsp_cnt.into());
-                    }
-                    others => {
-                        // NOTE: A call to `Client::call` implementation *MUST* always return the `Response` variant matching the `Request` one.
-                        // TIPS: This can be ensured via a call to `verify_response_header`( in 'src/service/mod.rs') before returning from `Client::call`.
-                        unreachable!(
-                            "{}",
-                            unexpected_rsp_code_panic_msg(
-                                FunctionCode::WriteMultipleRegisters,
-                                others.function_code()
-                            ),
-                        )
-                    }
-                })
+            .map(|result| {
+                result
+                    .map_err(Into::into)
+                    .and_then(|response| match response {
+                        Response::WriteMultipleRegisters(rsp_addr, rsp_cnt) => {
+                            debug_assert_eq!(addr, rsp_addr);
+                            debug_assert_eq!(cnt, rsp_cnt.into());
+                            Ok(())
+                        }
+                        response => Err(ResponseError::UnexpectedResponse { response }),
+                    })
             })
     }
 
@@ -395,31 +325,26 @@ impl Writer for Context {
         self.client
             .call(Request::MaskWriteRegister(addr, and_mask, or_mask))
             .await
-            .map(|modbus_rsp| {
-                modbus_rsp.map(|rsp| match rsp {
-                    Response::MaskWriteRegister(rsp_addr, rsp_and_mask, rsp_or_mask) => {
-                        debug_assert_eq!(addr, rsp_addr);
-                        debug_assert_eq!(and_mask, rsp_and_mask);
-                        debug_assert_eq!(or_mask, rsp_or_mask);
-                    }
-                    others => {
-                        // NOTE: A call to `Client::call` implementation *MUST* always return the `Response` variant matching the `Request` one.
-                        // TIPS: This can be ensured via a call to `verify_response_header`( in 'src/service/mod.rs') before returning from `Client::call`.
-                        unreachable!(
-                            "{}",
-                            unexpected_rsp_code_panic_msg(
-                                FunctionCode::MaskWriteRegister,
-                                others.function_code()
-                            ),
-                        )
-                    }
-                })
+            .map(|result| {
+                result
+                    .map_err(Into::into)
+                    .and_then(|response| match response {
+                        Response::MaskWriteRegister(rsp_addr, rsp_and_mask, rsp_or_mask) => {
+                            debug_assert_eq!(addr, rsp_addr);
+                            debug_assert_eq!(and_mask, rsp_and_mask);
+                            debug_assert_eq!(or_mask, rsp_or_mask);
+                            Ok(())
+                        }
+                        response => Err(ResponseError::UnexpectedResponse { response }),
+                    })
             })
     }
 }
 
 #[cfg(test)]
 mod tests {
+    use crate::ResponseResult;
+
     use super::*;
     use std::sync::Mutex;
 
@@ -427,7 +352,7 @@ mod tests {
     pub(crate) struct ClientMock {
         slave: Option<Slave>,
         last_request: Mutex<Option<Request<'static>>>,
-        next_response: Option<Result<Response>>,
+        next_response: Option<ResponseResult>,
     }
 
     #[allow(dead_code)]
@@ -440,17 +365,17 @@ mod tests {
             &self.last_request
         }
 
-        pub(crate) fn set_next_response(&mut self, next_response: Result<Response>) {
+        pub(crate) fn set_next_response(&mut self, next_response: ResponseResult) {
             self.next_response = Some(next_response);
         }
     }
 
     #[async_trait]
     impl Client for ClientMock {
-        async fn call(&mut self, request: Request<'_>) -> Result<Response> {
+        async fn call(&mut self, request: Request<'_>) -> ResponseResult {
             *self.last_request.lock().unwrap() = Some(request.into_owned());
-            match self.next_response.as_ref().unwrap() {
-                Ok(response) => Ok(response.clone()),
+            match self.next_response.take().unwrap() {
+                Ok(response) => Ok(response),
                 Err(err) => Err(io::Error::new(err.kind(), format!("{err}"))),
             }
         }