+++ tests/exception/mod.rs
@@ -6,7 +6,7 @@ use std::future;
 use tokio_modbus::{
     client::{Context, Reader, Writer},
     server::Service,
-    Exception, Request, Response,
+    Exception, Request, Response, ResponseError,
 };
 
 pub struct TestService {}
@@ -43,65 +43,99 @@ impl Service for TestService {
 // TODO: Update the `assert_eq` with a check on Exception once Client trait can return Exception
 pub async fn check_client_context(mut ctx: Context) {
     let response = ctx.read_coils(0x00, 2).await.expect("communication failed");
-    assert_eq!(response, Err(Exception::Acknowledge));
+    assert!(matches!(
+        response,
+        Err(ResponseError::Exception(Exception::Acknowledge))
+    ));
 
     let response = ctx
         .read_discrete_inputs(0x00, 2)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::GatewayPathUnavailable));
+    assert!(matches!(
+        response,
+        Err(ResponseError::Exception(Exception::GatewayPathUnavailable))
+    ));
 
     let response = ctx
         .write_single_coil(0x00, true)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::GatewayTargetDevice));
+    assert!(matches!(
+        response,
+        Err(ResponseError::Exception(Exception::GatewayTargetDevice))
+    ));
 
     let response = ctx
         .write_multiple_coils(0x00, &[true])
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::IllegalDataAddress));
+    assert!(matches!(
+        response,
+        Err(ResponseError::Exception(Exception::IllegalDataAddress))
+    ));
 
     let response = ctx
         .read_input_registers(0x00, 2)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::IllegalDataValue));
+    assert!(matches!(
+        response,
+        Err(ResponseError::Exception(Exception::IllegalDataValue))
+    ));
 
     let response = ctx
         .read_holding_registers(0x00, 2)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::IllegalFunction));
+    assert!(matches!(
+        response,
+        Err(ResponseError::Exception(Exception::IllegalFunction))
+    ));
 
     let response = ctx
         .write_single_register(0x00, 42)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::MemoryParityError));
+    assert!(matches!(
+        response,
+        Err(ResponseError::Exception(Exception::MemoryParityError))
+    ));
 
     let response = ctx
         .write_multiple_registers(0x00, &[42])
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::ServerDeviceBusy));
+    assert!(matches!(
+        response,
+        Err(ResponseError::Exception(Exception::ServerDeviceBusy))
+    ));
 
     let response = ctx
         .masked_write_register(0x00, 0, 0)
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::ServerDeviceFailure));
+    assert!(matches!(
+        response,
+        Err(ResponseError::Exception(Exception::ServerDeviceFailure))
+    ));
 
     let response = ctx
         .read_write_multiple_registers(0x00, 0, 0, &[42])
         .await
         .expect("communication failed");
-    assert_eq!(response, Err(Exception::IllegalFunction));
+    assert!(matches!(
+        response,
+        Err(ResponseError::Exception(Exception::IllegalFunction))
+    ));
 
     // TODO: This codes hangs if used with `rtu-over-tcp-server`, need to check why
-    /*
-    let response = ctx.call(Request::Custom(70, Cow::Owned(vec![42]))).await.expect("communication failed");
-    assert_eq!(response, Err(Exception::IllegalFunction));
-    */
+    // let response = ctx
+    //     .call(Request::Custom(70, Cow::Owned(vec![42])))
+    //     .await
+    //     .expect("communication failed");
+    // assert!(matches!(
+    //     response,
+    //     Err(ResponseError::Exception(Exception::IllegalFunction))
+    // ));
 }